Quan ly san pham va Quan ly don hang

+++ database/migrations/2025_04_15_140119_add_avatar_to_users_table.php
@@ -19,8 +19,11 @@ return new class extends Migration
     public function down(): void
     {
         Schema::table('users', function (Blueprint $table) {
-            $table->dropColumn('avatar');
+            if (Schema::hasColumn('users', 'avatar')) {
+                $table->dropColumn('avatar');
+            }
         });
     }
+
 
 };
